Fixed yet another bug in calcLockedShareAmount
